Implement an IK control in limbs.super_finger with snapping.
BlenRig has this feature, so Rigify was behind in this aspect.

Since this is a feature for specific situations, which applies
to 10 fingers if enabled, it is implemented in a lightweight
way using the minimal number of new bones; specifically, IK
is applied to the result of regular FK transformation, so
existing FK bones can be used to tweak the shape.

Another effect of that is that when IK is off, FK->IK snap
simply normalizes the finger by absorbing as much of rotation
as possible into the master control. This makes the operation
useful even if IK is not used or even generated.

For IK follow parent, it makes zero sense to follow the palm
bones, so the rig has an option for explicitly specifying the
parent bone. Natural choices are the hand, or a separate utility
bone for something held in the hand, if the metarig has it.