Port the Pivot Slide feature of the legacy Rigify spine to super_spine.
Pivot Slide seems to be the largest feature that was lost when merging
original and PitchiPoy rigs to form the new rig set. This ports it as
an option in the new spine rig, thus making the legacy spine completely
obsolete.

When not enabled, the generated rig shouldn't change. Also, head/tail
and tweaks should work the same even with the dynamic pivot mode.

The code is directly based on the legacy implementation, but extensively
edited to make it fit into super_spine code.