Add new utilities for constraint, property and driver creation.
The goal is to provide compact and easy to read & write syntax for
creating these entities. This can be achieved by exploiting functions
that can accept arbitrary keyword arguments, and flexible recursive
data structure parsing to initialize driver variables via nested
dictionaries and lists.

The spine rig is refactored to use the new utilities as an example.