Split and convert the super_spine rig to the new format.
Combining spine, head and tail in the same rig leads to overly
complicated and inflexible code. The only reason to do that is
to be able to join the B-Bone chains together, and with the new
rig class structure it can be done via rig interaction.

Thus the rig is split into a basic spine rig, and a head rig
that can optionally attach its deform chain to the parent.