Rigify: prepare to set custom property defaults and overridable flags.
In 2.8 custom property creation becomes more complex, with the
need to set a default value for correct NLA blending and an
overridable flag for static overrides. The most logical way
to do that is to refactor rigs to use the new make_property
utility, and implement the new features there.

This backports all those changes minus the actual new features.